Created new route <Search>
Installed helper that makes calls to the third party api and
caches the data.
Created a route that handles calls to the 3rd party api pokeapi.
It takes two params one is the path and the other is the name of
the object in that path.
Defined search action in the pokemon controller.
Created a script for the search action.